added hsv to rgb converter

+++ mosaic/ColorSpaces/Hsv.cs
@@ -2,16 +2,16 @@
 {
     internal struct Hsv
     {
-        public Hsv(int hue, int saturation, int value) : this()
+        public Hsv(decimal hue, decimal saturation, decimal value) : this()
         {
             H = hue;
             S = saturation;
             V = value;
         }
 
-        public int H { get; }
-        public int S { get; }
-        public int V { get; }
+        public decimal H { get; }
+        public decimal S { get; }
+        public decimal V { get; }
 
         public override string ToString()
         {
